Push 1
This is a basic layout I created for a company in my Sophomore summer
break. It is meant to be viewer of an oncological database. The code
here contains only a basic layout as the information was confidential.
It contains elements from HTML5, CSS3, JS, D3 and some jQuery packages.
Best viewed in Firefox.